+++ src/main/java/com/tagmycode/plugin/CrashService.java
@@ -1,5 +1,6 @@
 package com.tagmycode.plugin;
 
+import com.tagmycode.sdk.TagMyCode;
 import com.tagmycode.sdk.crash.Crash;
 import com.tagmycode.sdk.crash.CrashClient;
 import com.tagmycode.sdk.exception.TagMyCodeException;
@@ -10,12 +11,14 @@ import java.util.concurrent.Executors;
 public class CrashService {
     private CrashClient crashClient;
     private Data data;
+    private TagMyCode tagMyCode;
     private AbstractVersion version;
     private String appId;
 
-    public CrashService(CrashClient crashClient, Data data, AbstractVersion version, String consumerId) {
+    public CrashService(CrashClient crashClient, Data data, TagMyCode tagMyCode, AbstractVersion version, String consumerId) {
         this.crashClient = crashClient;
         this.data = data;
+        this.tagMyCode = tagMyCode;
         this.version = version;
         appId = consumerId;
     }
@@ -32,7 +35,7 @@ public class CrashService {
             public void run() {
                 try {
                     String appVersion = version.getFrameworkVersion() + " " + version.getPluginVersion() + " " + version.getPluginTitle();
-                    Crash crash = Crash.create(appId, appVersion, data.getAccount(), exception);
+                    Crash crash = Crash.create(appId, appVersion, data.getAccount(), tagMyCode.getClient().getOauthToken(), exception);
                     crashClient.sendCrash(crash);
                 } catch (TagMyCodeException tmcException) {
                     Framework.LOGGER.debug("Send crash exception", tmcException);